FirstTests
- Compatibility Changes for Python3
- Delete PySide references to use PyQt4 instead

+++ newProcessCorrelations.py
@@ -20,14 +20,14 @@ import os
 import initData
 
 def prepareCorrelations(fileNameList, gridX, gridY, corrsize, baseMode, floatStep, parentWidget, parentWindow, largeDisp, filterInfos, thread):
-    
+
     startTime = time.time()
     # getting the images filename list
     #fileName = parentWindow.fileDataPath+'/filenamelist.dat'
     #fileNameList = getData.testReadFile(fileName)
-    
-    infosThread = thread.signal.threadSignal    
-    
+
+    infosThread = thread.signal.threadSignal
+
     infosAnalysis = []
     isLargeDisp = 1
     # Name
@@ -40,26 +40,26 @@ def prepareCorrelations(fileNameList, gridX, gridY, corrsize, baseMode, floatSte
     # nbImages * nbMarkers
     # largeDisp YES/NO
     # User Profile
-    
+
     for image in range(len(fileNameList)):
         fileNameList[image] = fileNameList[image].rstrip()
     fileNameList = np.array(fileNameList)
-    
+
     activeImages = parentWidget.imageActiveList
-    
+
     if fileNameList is None:
         return
-        
+
     if largeDisp is None:
         largeDisp = np.zeros((len(fileNameList),2))
         isLargeDisp = 0
-    
-        
+
+
     numOfBasePoints = len(gridX)
     numOfImages = len(fileNameList)
 
-    #parentWidget.calculationBar.changeValue(1, 'Starting Processes...')        
-        
+    #parentWidget.calculationBar.changeValue(1, 'Starting Processes...')
+
     # Adding some informations to the DevMode widget
     parentWindow.devWindow.addInfo('Starting the correlation process with '+str(numOfBasePoints)+' markers on '+str(numOfImages)+' images.')
     if baseMode == 0:
@@ -68,8 +68,8 @@ def prepareCorrelations(fileNameList, gridX, gridY, corrsize, baseMode, floatSte
         parentWindow.devWindow.addInfo('Reference Image : First')
     elif baseMode == 2:
         parentWindow.devWindow.addInfo('Reference Image : Shifted ('+str(floatStep)+')')
-    
-    
+
+
     # Setting up the processes
     PROCESSES = int(parentWindow.profileData['nbProcesses'][parentWindow.currentProfile])
     args = []
@@ -79,28 +79,28 @@ def prepareCorrelations(fileNameList, gridX, gridY, corrsize, baseMode, floatSte
         PROCESSES = numOfBasePoints/2
     parentWindow.devWindow.addInfo('Number of processes used: '+str(PROCESSES))
     for i in range(0,PROCESSES):
-        start = i*nbMarkersPerProcess
+        start = int(i*nbMarkersPerProcess)
         if i >= PROCESSES-1: #last process do all the last images
             end = numOfBasePoints
         else:
-            end = (i+1)*nbMarkersPerProcess 
+            end = int((i+1)*nbMarkersPerProcess)
         args.append((fileNameList, activeImages, parentWindow.filePath, gridX[start:end], gridY[start:end], baseMode, corrsize, floatStep, largeDisp, filterInfos))
 
     result = parentWindow.createProcess(processCorrelation, args, PROCESSES, parentWidget.calculationBar, '(1/2) Processing images...')
-    
-    
+
+
     parentWindow.devWindow.addInfo('Calculation finished. Saving data files.')
 
     #neighbors calculation
-    parentWidget.calculationBar.changeValue(1, '(2/2) Calculating neighborhood...') 
-    activeMarkers = np.linspace(0, numOfBasePoints, num=numOfBasePoints, endpoint=False)
+    parentWidget.calculationBar.changeValue(1, '(2/2) Calculating neighborhood...')
+    activeMarkers = np.linspace(0, numOfBasePoints, num=numOfBasePoints, endpoint=False).astype(np.int)
     minNeighbors = 16
     maxCorrDistance = 100
     initData.calculateNeighbors(activeMarkers, result[0][:,image], result[1][:,image], minNeighbors, maxCorrDistance, parentWindow.fileDataPath, progressBar=parentWidget.calculationBar)
-    
+
 
     #data saving
-    parentWidget.calculationBar.changeValue(100, 'Saving .dat files...')  
+    parentWidget.calculationBar.changeValue(100, 'Saving .dat files...')
     Save('validx.dat', result[0], parentWindow.fileDataPath)
     Save('validy.dat', result[1], parentWindow.fileDataPath)
     Save('stdx.dat', result[3], parentWindow.fileDataPath)
@@ -115,7 +115,7 @@ def prepareCorrelations(fileNameList, gridX, gridY, corrsize, baseMode, floatSte
 
     parentWindow.devWindow.addInfo('Calculation done. Data files saved.')
     totalTime = time.time() - startTime
-    
+
     infosAnalysis.append(os.path.basename(parentWindow.fileDataPath))
     infosAnalysis.append(baseMode)
     infosAnalysis.append(corrsize)
@@ -129,21 +129,21 @@ def prepareCorrelations(fileNameList, gridX, gridY, corrsize, baseMode, floatSte
     Save('infoAnalysis.dat', np.array(infosAnalysis), parentWindow.fileDataPath)
     if isLargeDisp:
         Save('largeDisp.dat', largeDisp, parentWindow.fileDataPath)
-    
-    
+
+
     parentWindow.devWindow.addInfo('Processing Time : '+str(totalTime))
     infosThread.emit([1])
     return
 
 
 def processCorrelation(fileNameList, activeImages, filePath, gridX, gridY, baseMode, corrsize, floatStep, largeDisp, filterInfos, q, pipe):
-    
-    
+
+
     # Initialise variables:
     [basePointsX, basePointsY, inputPointsX, inputPointsY] = InitFunc(gridX, gridY)
     numOfBasePoints = len(basePointsX)
     numOfImages = len(fileNameList)
-    
+
     # Creating the main matrices to register the data
     ValidX = np.zeros((numOfBasePoints, numOfImages))
     ValidY = np.zeros((numOfBasePoints, numOfImages))
@@ -153,10 +153,10 @@ def processCorrelation(fileNameList, activeImages, filePath, gridX, gridY, baseM
     DispX = np.zeros((numOfBasePoints,numOfImages))
     DispY = np.zeros((numOfBasePoints,numOfImages))
     infoMarkers = np.zeros((numOfBasePoints,numOfImages))
-    
-    
+
+
     result = np.zeros((8, numOfBasePoints, numOfImages))
-    
+
     refImg = 0
     # Loading the reference image and applying filter if exist
     while(activeImages[refImg] == 0):
@@ -170,33 +170,33 @@ def processCorrelation(fileNameList, activeImages, filePath, gridX, gridY, baseM
         refImg += 1
     base = cv2.imread(filePath+'/'+fileNameList[refImg], 0) # Reference image
     #base = cv2.cvtColor(baseRaw, cv2.COLOR_BGR2GRAY)
-    
+
 
     #apply filter if loaded
     base = filterWidget.applyFilterListToImage(filterInfos, base)
-                    
-                    
+
+
     ValidX[:,refImg]=basePointsX[:,0]
     ValidY[:,refImg]=basePointsY[:,0]
-    
-        
-    previousTime = time.time()   
+
+
+    previousTime = time.time()
     # Process all images: calculate correlation between reference and current image
     for CurrentImage in range(refImg+1, numOfImages):
-        
-        
+
+
         currentProgress = CurrentImage * 100 / numOfImages
         currentTime = time.time()
         if currentTime > previousTime + .05:
             previousTime = currentTime
-            pipe.send(currentProgress)      
-        
+            pipe.send(currentProgress)
+
         if activeImages[CurrentImage] == 1:
-        
+
             inputImg = cv2.imread(filePath+'/'+fileNameList[CurrentImage], 0)
             inputImg = filterWidget.applyFilterListToImage(filterInfos, inputImg)
             #inputImg = cv2.cvtColor(inputRaw, cv2.COLOR_BGR2GRAY)
-            
+
             if baseMode == 2:
                 #Reference image is shifted from float
                 if sum(activeImages[0:CurrentImage]) > floatStep:
@@ -209,23 +209,23 @@ def processCorrelation(fileNameList, activeImages, filePath, gridX, gridY, baseM
                     newY = ValidY[:,CurrentImage-floatStep-imageSelection]
                     basePointsX = np.reshape(newX, (len(newX),1))
                     basePointsY = np.reshape(newY, (len(newY),1))
-    
-            
-            
+
+
+
             #adding large displacement
             largeDisplacementX = largeDisp[CurrentImage, 0] - largeDisp[CurrentImage-1, 0]
             largeDisplacementY = largeDisp[CurrentImage, 1] - largeDisp[CurrentImage-1, 1]
-                
-    
+
+
             inputPointsX = inputPointsX + largeDisplacementX
             inputPointsY = inputPointsY + largeDisplacementY
-            
-            
+
+
             [inputCorrX, inputCorrY, currentStdX, currentStdY, currentCorrCoef, infosError] = CpcorrFunc(basePointsX, basePointsY, inputPointsX, inputPointsY, base, inputImg, corrsize)
             inputPointsX = inputCorrX
-            inputPointsY = inputCorrY        
-            
-            
+            inputPointsY = inputCorrY
+
+
             if baseMode == 0:
                 # Reference is previous image
                 base = inputImg
@@ -233,8 +233,8 @@ def processCorrelation(fileNameList, activeImages, filePath, gridX, gridY, baseM
                 basePointsY = inputCorrY
                 #basePointsX = np.round(inputCorrX, decimals = 1)
                 #basePointsY = np.round(inputCorrY, decimals = 1)
-                
-            
+
+
             [ValidXCollected,ValidYCollected,StdXCollected,StdYCollected,CorrCoefCollected] = CollectDataFunc(inputCorrX, inputCorrY, currentStdX, currentStdY, currentCorrCoef) # a column for each image
             #[self.XSize, self.YSize] = self.input.shape
             ValidX[:,CurrentImage] = ValidXCollected[:,0]
@@ -242,12 +242,12 @@ def processCorrelation(fileNameList, activeImages, filePath, gridX, gridY, baseM
             DispX[:, CurrentImage] = ValidX[:, CurrentImage] - ValidX[:, refImg] - largeDisp[CurrentImage, 0]
             DispY[:, CurrentImage] = ValidY[:, CurrentImage] - ValidY[:, refImg] - largeDisp[CurrentImage, 1]
             CorrCoef[:,CurrentImage] = CorrCoefCollected[:,0]
-            StdX[:,CurrentImage] = StdXCollected[:,0] 
+            StdX[:,CurrentImage] = StdXCollected[:,0]
             StdY[:,CurrentImage] = StdYCollected[:,0]
             infoMarkers[:,CurrentImage] = infosError[:,0]
-            
+
         else: #image has been removed by the user, putting NaN values in data
-        
+
             ValidX[:,CurrentImage] = np.nan
             ValidY[:,CurrentImage] = np.nan
             DispX[:, CurrentImage] = np.nan
@@ -255,9 +255,9 @@ def processCorrelation(fileNameList, activeImages, filePath, gridX, gridY, baseM
             CorrCoef[:,CurrentImage] = np.nan
             StdX[:,CurrentImage] = np.nan
             StdY[:,CurrentImage] = np.nan
-            
-        
-        
+
+
+
     #assembling all matrices into one 3D result matrice
     result[0] = ValidX
     result[1] = ValidY
@@ -267,21 +267,21 @@ def processCorrelation(fileNameList, activeImages, filePath, gridX, gridY, baseM
     result[5] = DispX
     result[6] = DispY
     result[7] = infoMarkers
-    
+
     q.put(result)
     q.close()
     return
-        
-        
+
+
 
 def InitFunc(GridX, GridY):
     # Initialize variables
-    inputPointsX = np.reshape(GridX,(len(GridX),1)) 
+    inputPointsX = np.reshape(GridX,(len(GridX),1))
     inputPointsY = np.reshape(GridY,(len(GridY),1))
     basePointsX = np.reshape(GridX,(len(GridX),1))
     basePointsY = np.reshape(GridY,(len(GridY),1))
     return basePointsX, basePointsY, inputPointsX, inputPointsY
-    
+
 
 def CpcorrFunc(BasePointsX, BasePointsY, InputPointsX, InputPointsY, Base, Input, corrsize):
 
@@ -312,21 +312,21 @@ def CollectDataFunc(InputCorrX,InputCorrY,StdX,StdY,CorrCoef):
     stdXCurrent = StdX
     stdYCurrent = StdY
     return validXCurrent, validYCurrent, stdXCurrent, stdYCurrent, corrCoefCurrent
-    
+
 def Save(FileName, Data, filePath):
 
     np.savetxt(filePath+'/'+FileName, Data, fmt="%s")
 
-    
+
 def shiftDetection(filePath, imageList, activeImages, area, filterList, thread):
-    
+
     largeDisp = np.zeros((len(imageList),2))
 
     initImage = cv2.imread(filePath+'/'+imageList[0].rstrip(), 0) #read the full image
     initImage = filterWidget.applyFilterListToImage(filterList, initImage)
     nbImages = len(imageList)
     currentPercent = 1
-    
+
     activeFileList = []
     for image in range(1, nbImages):
         if activeImages[image] == 1:
@@ -334,22 +334,22 @@ def shiftDetection(filePath, imageList, activeImages, area, filterList, thread):
 
     template = initImage[area[1]:area[3],area[0]:area[2]] #select the template data
     width = area[2]-area[0]
-    height = area[3]-area[1]   
+    height = area[3]-area[1]
 
     origin = (area[0], area[1])
-    startTime = time.time() 
+    startTime = time.time()
     for i in activeFileList:
-        
+
         newImage = cv2.imread(filePath+'/'+imageList[i].rstrip(),0)
         newImage = filterWidget.applyFilterListToImage(filterList, newImage)
 
         matchArea = cv2.matchTemplate(newImage, template, cv2.TM_CCORR_NORMED)
         minVal, maxVal, minLoc, maxLoc = cv2.minMaxLoc(matchArea)
         template = newImage[maxLoc[1]:maxLoc[1]+height,maxLoc[0]:maxLoc[0]+width] #the template for the next image is update with the template found on the current picture
-        
+
         largeDisp[i][0] = maxLoc[0]-origin[0] #save the displacement
         largeDisp[i][1] = maxLoc[1]-origin[1]
-        
+
         percent = i*100/nbImages
         if percent > currentPercent:
             thread.signal.threadSignal.emit([percent, i, largeDisp[i][0], largeDisp[i][1]])
@@ -358,4 +358,3 @@ def shiftDetection(filePath, imageList, activeImages, area, filterList, thread):
     totalTime = time.time() - startTime
     thread.signal.threadSignal.emit([100, nbImages, largeDisp, totalTime])
     #print totalTime
-
